requirements.txt
Flask==2.3.4
requests==2.31.0
azure-identity==1.12.0
azure-keyvault-secrets==4.11.0